requirements.txt
app==0.0.1
discord_webhook==1.3.0
fastapi~=0.104.1
fastapi_versioning==0.10.0
passlib==1.7.4
pydantic==2.4.2
pymongo==4.5.0
python-dotenv==1.0.0
python_jose==3.3.0
typing_extensions==4.8.0
uvicorn==0.23.2
web3==6.11.1
python-multipart
pyotp~=2.9.0
asyncstdlib~=3.10.9
solana~=0.30.2